+++ mPak/Binaries/Win64/UnrealEditor.modules
@@ -1,0 +1,7 @@
+{
+	"BuildId": "25360045",
+	"Modules": 
+	{
+		"mPak": "UnrealEditor-mPak.dll"
+	}
+}
