+++ mPak/Source/mPak/Private/SMakeFileEd.cpp
@@ -1,0 +1,393 @@
+// Fill out your copyright notice in the Description page of Project Settings.
+
+
+#include "SMakeFileEd.h"
+#include "SlateOptMacros.h"
+#include <Kismet/KismetSystemLibrary.h>
+//#include "mPakCore.h"
+#include <UATHelper/Public/IUATHelperModule.h>
+#include <ContentBrowserModule.h>
+#include <IContentBrowserSingleton.h>
+
+
+
+BEGIN_SLATE_FUNCTION_BUILD_OPTIMIZATION
+
+#define LOCTEXT_NAMESPACE "SMakeFileEd"
+
+void SMakeFileEd::Construct(const FArguments& InArgs)
+{
+ 
+
+	ChildSlot
+		[
+			SNew(SBox)
+	.HAlign(HAlign_Center)
+	.VAlign(VAlign_Center)
+	[
+		SNew(SVerticalBox)
+		+ SVerticalBox::Slot()
+		.VAlign(VAlign_Center)
+		.HAlign(HAlign_Center)
+		.AutoHeight()
+		[
+
+			SNew(SHorizontalBox)
+			+ SHorizontalBox::Slot()
+			.VAlign(VAlign_Center)
+			.AutoWidth()
+		[
+			SNew(STextBlock)
+			.Text(LOCTEXT("a", "要打包的路径 "))
+		]
+	+ SHorizontalBox::Slot()
+		[
+			SNew(SBox)
+			.WidthOverride(300.f)
+		[
+
+			//SNew(SEditableTextBox)
+			SAssignNew(MakeEditableTextBox,SEditableTextBox)
+			.Text(this,&SMakeFileEd::GetSelectDirectory)
+
+		]
+		]
+
+
+		]
+
+		+SVerticalBox::Slot()
+			.VAlign(VAlign_Center)
+			.HAlign(HAlign_Center)
+			.AutoHeight()
+			[
+				SNew(SHorizontalBox)
+					+ SHorizontalBox::Slot()
+					.VAlign(VAlign_Center)
+					.AutoWidth()
+					[
+					SNew(STextBlock)
+					.Text(LOCTEXT("c", "要保存的路径 "))
+					]
+					+ SHorizontalBox::Slot()
+					[
+						SNew(SBox)
+						.WidthOverride(300.f)
+						[
+							//SNew(SEditableTextBox)
+							SAssignNew(SaveEditableTextBox, SEditableTextBox)
+							.Text(LOCTEXT("outFile", "C:/Users/hotWin/Desktop/car.mpak"))
+						]
+					]
+			]
+
+
+		+SVerticalBox::Slot()
+			.VAlign(VAlign_Center)
+			.HAlign(HAlign_Center)
+			.AutoHeight()
+				[
+					SNew(SHorizontalBox)
+					+ SHorizontalBox::Slot()
+					[
+						SNew(SVerticalBox)
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(WindowsEditorCheckBox,SCheckBox)
+								.IsChecked(true)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("WindowsEditor", "WindowsEditor"))
+								]
+							]
+						
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(WindowsCheckBox, SCheckBox)
+								.IsChecked(true)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("Windows", "Windows"))
+								]
+							]
+							
+							
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(HololensCheckBox, SCheckBox)
+								.IsEnabled(false)							
+								.IsChecked(false)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("Hololens", "Hololens"))
+								]
+							]
+
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(AndroidCheckBox, SCheckBox)
+								.IsEnabled(false)
+								.IsChecked(false)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("Android", "Android"))
+								]
+							]
+							
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(LinuxCheckBox, SCheckBox)
+								.IsEnabled(false)
+								.IsChecked(false)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("Linux", "Linux"))
+								]
+							]
+					
+						+ SVerticalBox::Slot()
+							[
+								//SNew(SCheckBox)
+								SAssignNew(IOSCheckBox, SCheckBox)
+								.IsEnabled(false)
+								.IsChecked(false)
+								[
+									SNew(STextBlock)
+									.Text(LOCTEXT("IOS", "IOS"))
+								]
+							]
+					]
+
+					+ SHorizontalBox::Slot()
+					.HAlign(HAlign_Right)
+					.VAlign(VAlign_Bottom)
+					[
+						SNew(SBox)
+						.WidthOverride(270.f)
+						[
+						
+							SAssignNew(PackButton,SButton)
+							.Text(LOCTEXT("打包", "打包"))	
+							.OnClicked_Raw(this, &SMakeFileEd::OnPackButtonClickFun)
+							.HAlign(HAlign_Center)
+							.VAlign(VAlign_Center)
+						]
+					]
+		   		]
+			]
+		];
+
+}
+
+END_SLATE_FUNCTION_BUILD_OPTIMIZATION
+
+bool SMakeFileEd::MakePathAndCommand()
+{
+
+	//UEAccessPath
+	UEAccess_Path = MakeEditableTextBox->GetText().ToString() + "/*";
+
+	//mpak file out path
+	mPakOut_Path = SaveEditableTextBox->GetText().ToString();
+
+
+	//ProjectRoot
+	int32 ContentIndex = MakeEditableTextBox->GetText().ToString().Find(TEXT("Content"));
+	ProjectRoot = MakeEditableTextBox->GetText().ToString().Left(ContentIndex - 1);
+
+
+	//ProjectName
+	ProjectName = UKismetSystemLibrary::GetGameName();
+
+	//Project_uproject
+	//Project_uproject = ProjectRoot + "/" + ProjectName + ".uproject";
+	Project_uproject = FPaths::ConvertRelativePathToFull(FPaths::GetProjectFilePath());
+
+	//CookRight
+	CookRight = MakeEditableTextBox->GetText().ToString().Right(MakeEditableTextBox->GetText().ToString().Len() - ProjectRoot.Len());
+
+
+	//打包编辑器命令
+	PackPieCmd = "UnrealPak " + pakTempDir + "/pie.pak " + UEAccess_Path;
+
+	//打包Windows包命令
+	PackWindowsCmd = "UnrealPak " + pakTempDir + "/run.pak " + ProjectRoot + "/Saved/Cooked/Windows/" + ProjectName + CookRight + "/*";
+	
+	
+	return true;
+
+}
+
+
+
+void SMakeFileEd::DoPackFun()
+{
+	FString EngineDir = FPaths::EngineDir();
+	FString UnrealEditorCmdPath = FPaths::Combine(*EngineDir, TEXT("Binaries"), TEXT("Win64"), TEXT("UnrealEditor-Cmd.exe"));
+
+
+	FString  str1 = UnrealEditorCmdPath + " ";
+	FString  str2 = Project_uproject;
+	FString  str3 = " -run=Cook -TargetPlatform=Windows -unversioned -stdout -CrashForUAT -unattended -NoLogTimes -UTF8Output";
+   	FString cookWindowsCmd = str1 + str2 + str3;
+
+
+
+	//引擎的转换方法会出问题
+	std::string MyStdString1(TCHAR_TO_UTF8(*cookWindowsCmd));
+	char cookCommand[] = "";
+	MyStdString1.copy(cookCommand, MyStdString1.size(), 0);
+
+
+	const char* packPieCommand = TCHAR_TO_ANSI(*PackPieCmd);		//UnrealPak C:/Users/hotPC/Desktop/editP/Saved/Sandboxes/pie.pak C:/Users/hotPC/Desktop/editP/Content/makePak/*
+
+
+
+	std::string MyStdString(TCHAR_TO_UTF8(*PackWindowsCmd));
+	char PackWinCommand[]="";
+	MyStdString.copy(PackWinCommand, MyStdString.size(), 0);
+
+
+	
+	FString ProjectPluginsDir = FPaths::ConvertRelativePathToFull(FPaths::ProjectPluginsDir());
+	FString Batpath = ProjectPluginsDir += "mPakImporter_UE5/ThirdParty/BuildMPak.bat";
+	
+
+
+
+	FPaths::NormalizeFilename(Batpath);
+
+
+
+	//::ProjectName
+	//::TargetPakFile_pie
+	//::TargetPakFile_run
+	//::PakSourceListFile_pie
+	//::PakSourceListFile_run
+	//::mPakOutput
+  	FString TargetPakFile_pie = FString(pakTempDir + "/pie.pak");
+	FString TargetPakFile_run = FString(pakTempDir + "/run.pak");
+	FString PakSourceListFile_pie = FString(pakTempDir + "/run.pak");
+	FString PakSourceListFile_run = FString(ProjectRoot + "/Saved/Cooked/Windows/" + ProjectName + CookRight + "/*");
+
+
+
+
+	FString batParam;
+	batParam += FString::Printf(TEXT(" %s"), *Project_uproject);
+	batParam += FString::Printf(TEXT(" %s"), *TargetPakFile_pie);
+	batParam += FString::Printf(TEXT(" %s"), *TargetPakFile_run);
+	batParam += FString::Printf(TEXT(" %s"), *UEAccess_Path);
+	batParam += FString::Printf(TEXT(" %s"), *PakSourceListFile_run);
+	batParam += FString::Printf(TEXT(" %s"), *mPakOut_Path);
+
+
+
+
+
+	//启动器是否隐藏
+	//启动器是否真正的隐藏	   3
+	//返回创建进程的ID
+	//优先级默认0
+	//指定工作路径
+	//剩下两个参数是管线相关的内容
+	FProcHandle Handle = FPlatformProcess::CreateProc(*Batpath, *batParam, false, false, false, nullptr, 0, NULL, nullptr, nullptr);
+	FPlatformProcess::WaitForProc(Handle);
+
+}
+
+
+
+void SMakeFileEd::CookWin()
+{
+    //获取各种路径
+	FString ProjectPath = FPaths::ConvertRelativePathToFull(FPaths::ProjectDir());
+
+	FString FullProjectName = FPaths::ConvertRelativePathToFull(FPaths::GetProjectFilePath());
+
+	FString EngineDir = FPaths::ConvertRelativePathToFull(FPaths::EngineDir());
+	FString EditorCmdPath = FPaths::Combine(*EngineDir, TEXT("Binaries"), TEXT("Win64"), TEXT("UnrealEditor-Cmd.exe"));	 \
+
+	FString ProjectPluginsDir = FPaths::ConvertRelativePathToFull(FPaths::ProjectPluginsDir());
+	FString cookBatpath = ProjectPluginsDir += "mPak/ThirdParty/CookContent.bat";
+
+	FString TargetPlatform = TEXT("Windows");
+
+	
+
+	//路径有可能存在空格
+	FString batParam;
+	batParam += FString::Printf(TEXT(" \"%s\""), *EditorCmdPath);
+	batParam += FString::Printf(TEXT(" \"%s\""), *FullProjectName);
+	batParam += FString::Printf(TEXT(" \"%s\""), *TargetPlatform);
+
+  	/*
+	C:/Program Files/Epic Games/UE_5.2/Engine/Binaries/Win64/UnrealEditor-Cmd.exe
+	C:/Users/hotWin/Desktop/dddd/dddd.uproject
+	Windows
+	*/
+
+
+	FProcHandle Handle = FPlatformProcess::CreateProc(*cookBatpath, *batParam, false, false, false, nullptr, 0, NULL, nullptr, nullptr);
+	FPlatformProcess::WaitForProc(Handle);
+
+
+	UE_LOG(LogTemp, Warning, TEXT("%s"), *EditorCmdPath);
+	UE_LOG(LogTemp, Warning, TEXT("%s"), *FullProjectName);
+	UE_LOG(LogTemp, Warning, TEXT("%s"), *TargetPlatform);
+
+
+
+}
+FReply SMakeFileEd::OnPackButtonClickFun()
+{
+
+	CookWin();
+	//if (MakePathAndCommand())
+	//{
+	//	DoPackFun();
+	//	
+	//}
+
+  	return FReply::Handled();
+}
+
+
+
+FText SMakeFileEd::GetSelectDirectory() const
+{
+
+
+	FContentBrowserModule& ContentBrowserModule = FModuleManager::LoadModuleChecked<FContentBrowserModule>("ContentBrowser");
+	IContentBrowserSingleton& ContentBrowserSingleton = ContentBrowserModule.Get();
+	TArray<FString> Folders;
+	ContentBrowserSingleton.GetSelectedFolders(Folders);
+
+	
+	FString absContentPath = FPaths::ConvertRelativePathToFull(FPaths::ProjectDir()) + "Content";
+
+	if (Folders.Num() > 0)
+	{
+		FString FullSelectPaht = Folders[0].Replace(TEXT("/All/Game"), *absContentPath);
+		
+		//暂不支持插件路径
+		//FString FullSelectPaht = Folders[0].Replace(TEXT("/All/Plugins"), *absContentPath);
+
+		const FText res = FText::FromString(FullSelectPaht);
+		return res;
+	}
+	else
+	{
+		const FText res = FText::FromString("Please Select a folder in the Content browser");
+		return res;
+		
+	}
+
+}
+
+#undef LOCTEXT_NAMESPACE
