+++ mPak/mPak.uplugin
@@ -1,0 +1,24 @@
+{
+	"FileVersion": 3,
+	"Version": 1,
+	"VersionName": "1.0",
+	"FriendlyName": "mPak",
+	"Description": "",
+	"Category": "Other",
+	"CreatedBy": "hotMonk",
+	"CreatedByURL": "https://inite.cn",
+	"DocsURL": "",
+	"MarketplaceURL": "",
+	"SupportURL": "",
+	"CanContainContent": false,
+	"IsBetaVersion": true,
+	"IsExperimentalVersion": false,
+	"Installed": false,
+	"Modules": [
+		{
+			"Name": "mPak",
+			"Type": "Editor",
+			"LoadingPhase": "Default"
+		}
+	]
+}
