--- mPakImporter_UE5/Source/mPakImporter/Public/mPakImporter.h
@@ -1,14 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#pragma once
-
-#include "Modules/ModuleManager.h"
-
-class FmPakImporterModule : public IModuleInterface
-{
-public:
-
-	/** IModuleInterface implementation */
-	virtual void StartupModule() override;
-	virtual void ShutdownModule() override;
-};
