--- mPakImporter_UE5/mPak.uplugin
@@ -1,28 +1,0 @@
-{
-	"FileVersion": 3,
-	"Version": 1,
-	"VersionName": "1.0",
-	"FriendlyName": "mPak",
-	"Description": "Make and importer mPak file",
-	"Category": "Other",
-	"CreatedBy": "hotMonk",
-	"CreatedByURL": "http://www.inite.cn",
-	"DocsURL": "",
-	"MarketplaceURL": "",
-	"SupportURL": "",
-	"CanContainContent": false,
-	"IsBetaVersion": true,
-	"IsExperimentalVersion": false,
-	"Installed": false,
-	"Modules": [
-		{
-			"Name": "mPak",
-			"Type": "Editor",
-			"LoadingPhase": "Default"
-		},{
-			"Name": "mPakImporter",
-			"Type": "Runtime",
-			"LoadingPhase": "PreLoadingScreen"
-		}
-	]
-}
